Bayesian Multivariate Normal Modelling for Common Source Testing

Bayesian Multivariate Normal Modelling for Common Source Testing

Usage

BayesMuCoSoT_fit(
  y,
  x = NA,
  questioned_data,
  known_data,
  background_data = NA,
  background_data_id = NA,
  approach = "independent prior",
  BayesFactor_Approximation = "Bridge Sampling",
  prior_elicitation = "Maximum Likelihood",
  DoF = "min",
  return_jags_outputs = FALSE,
  verbose = TRUE,
  ...
)

Arguments

y

Vector of response columns names (p=variables).

x

Predictors columns names (d=predictors). If there are any. Including constant (all 1's vector)

questioned_data

Data frame with the unknown data.

known_data

Data frame of the known data.

background_data

Background data frame to estimate the prior parameters. Background data recommended.

background_data_id

Column name of the background data frame for discriminate the data per id.

approach

"conjugate" or "independent prior" approach. Default "independent prior"

BayesFactor_Approximation

"Laplace-Metropolis" or "Generalized Harmonic Mean" or "Bridge Sampling". Default "Bridge Sampling". For the conjugate approach the Bayes factor comes in close form.

prior_elicitation

"Non-Informative" or "Maximum Likelihood" or list of specific names and type list(U = the inverse Wishart's scale matrix p by p, beta_mu = matrix mean of the matrix Normal d by p, beta_cov = d covariances matrix (3D matrix p by p by d) or in case of conjugate approach beta_cov = prior measurements on the W scale matrix d by d. Default "Maximum Likelihood" estimations from the Background data.

DoF

Degrees of freedom for the prior inverse Wishart distribution. Default "min": p+2. Otherwise, integer bigger than p+2.

return_jags_outputs

default FALSE

verbose

messages while running. Default TRUE.

...

the other arguments of jags R2jags function.

Value

the logarithmic Bayes factor.

Examples

# example code for more examples see documentation in github
all_data = iris[iris$Species=='setosa',]
questioned_data = all_data[1:(nrow(all_data)/2),]
known_data = all_data[(nrow(all_data)/2+1):nrow(all_data),]
background_data = iris[iris$Species!='setosa',]
background_data_id = 'Species'
y = names(questioned_data)[1:4]
BayesMuCoSoT_fit(y, x=NA, questioned_data, known_data, background_data,background_data_id)
#> [1] "Prior elicitation finished"
#> [1] "Sampling from posterior considering common modeling per source"
#> [1] "Sampling from posterior considering different modeling per source"
#> [1] "Calculating Bayes factor"
#> [1] 19.76335